Refactored lib/application.rb
