+++ hailstorm-gem/features/support/feature_helper.rb
@@ -1,8 +1,10 @@
 require 'rubygems'
 require 'bundler/setup'
 require 'simplecov'
-require 'hailstorm/application'
 
 $LOAD_PATH.push(File.expand_path('../../lib', __FILE__))
 $CLASSPATH << File.expand_path('../../data', __FILE__)
 ENV['HAILSTORM_ENV'] = 'cucumber'
+
+require 'hailstorm/initializer'
+require 'hailstorm/initializer/eager_load'
