+++ hailstorm-gem/lib/hailstorm/cli.rb
@@ -1,0 +1,5 @@
+require 'hailstorm'
+
+# CLI module
+module Hailstorm::Cli
+end
