+++ hailstorm-gem/lib/hailstorm/controller.rb
@@ -1,0 +1,5 @@
+require 'hailstorm'
+
+# Namespace for Hailstorm controller classes
+module Hailstorm::Controller
+end
