+++ hailstorm-gem/lib/hailstorm/initializer/eager_load.rb
@@ -1,0 +1,9 @@
+require 'ostruct'
+require 'active_support/all'
+require 'active_record'
+require 'action_dispatch/http/mime_type'
+require 'action_view'
+
+require 'hailstorm/version'
+
+ActiveRecord::Base.raise_in_transactional_callbacks = true
