+++ hailstorm-gem/lib/hailstorm/initializer/tmp_directory.rb
@@ -1,0 +1,6 @@
+require 'hailstorm'
+
+# clear project's tmp dir
+Dir["#{Hailstorm.tmp_path}/*"].each do |e|
+  File.directory?(e) ? FileUtils.rmtree(e) : File.unlink(e)
+end
