+++ hailstorm-gem/lib/hailstorm/model/target_host.rb
@@ -188,10 +188,4 @@ class Hailstorm::Model::TargetHost < ActiveRecord::Base
       self.project.current_execution_cycle.id
     )
   end
-
-  private
-
-  def command
-    Hailstorm.application.command_processor
-  end
 end
