+++ hailstorm-gem/lib/hailstorm/version.rb
@@ -1,3 +1,3 @@
 module Hailstorm
-  VERSION = '4.0.0'
+  VERSION = '5.0.0'
 end
